backend

+++ ep-domain/src/main/java/com/ep/domain/service/MemberChildCommentService.java
@@ -6,7 +6,6 @@ import com.ep.domain.constant.BizConstant;
 import com.ep.domain.constant.MessageCode;
 import com.ep.domain.pojo.ResultDo;
 import com.ep.domain.pojo.bo.MemberChildCommentBo;
-import com.ep.domain.pojo.dto.OrganClassScheduleDto;
 import com.ep.domain.pojo.po.EpFilePo;
 import com.ep.domain.pojo.po.EpMemberChildCommentPo;
 import com.ep.domain.pojo.po.EpMemberChildTagPo;
@@ -15,14 +14,12 @@ import com.ep.domain.repository.*;
 import com.ep.domain.repository.domain.enums.EpMemberChildCommentType;
 import com.google.common.collect.Lists;
 import lombok.extern.slf4j.Slf4j;
-import org.jooq.Condition;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 
@@ -114,22 +111,6 @@ public class MemberChildCommentService {
         return ResultDo.build();
     }
 
-    /**
-     * 商户后台获取分页
-     *
-     * @param pageable
-     * @param conditions
-     * @return
-     */
-    public Page<OrganClassScheduleDto> findbyPageAndCondition(Long courseId, Pageable pageable, Collection<Condition> conditions) {
-        Page<OrganClassScheduleDto> page = memberChildCommentRepository.findbyPageAndCondition(pageable, conditions);
-        page.getContent().forEach(p -> {
-            if (p.getCourseId() == null) {
-                p.setCourseId(courseId);
-            }
-        });
-        return page;
-    }
 
     /**
      * 商户后台修改评论内容
@@ -174,16 +155,19 @@ public class MemberChildCommentService {
                 insertPos.add(po);
             });
         }
-        memberChildCommentRepository.updateContent(id, content);
-        //先物理删除孩子标签，再插入
-        memberChildTagRepository.deletePhysicByChildIdAndClassCatalogId(childId, classCatalogId);
-        memberChildTagRepository.insert(insertPos);
-        log.info("[评论]修改评论成功，评论id={}。");
-        return ResultDo.build();
+        if (memberChildCommentRepository.updateContent(id, content) == BizConstant.DB_NUM_ONE) {
+            //先物理删除孩子标签，再插入
+            memberChildTagRepository.deletePhysicByChildIdAndClassCatalogId(childId, classCatalogId);
+            memberChildTagRepository.insert(insertPos);
+            log.info("[评论]修改评论成功，评论id={}。");
+            return ResultDo.build();
+        } else {
+            return ResultDo.build(MessageCode.ERROR_OPERATE_FAIL);
+        }
     }
 
     /**
-     * 后台创建随堂评论
+     * 由教师后台创建随堂评论
      *
      * @param childId
      * @param ognId
@@ -194,25 +178,86 @@ public class MemberChildCommentService {
      * @param mobile
      * @return
      */
-    public ResultDo createCommentLaunch(Long childId, Long ognId, Long courseId, Long classId, Long catalogId, String content, Long mobile) {
+    @Transactional(rollbackFor = Exception.class)
+    public ResultDo createCommentLaunchByAccount(Long childId, Long ognId, Long courseId, Long classId, Long catalogId, String content, Long mobile, List<Long> tagIds) {
         Optional<EpOrganAccountPo> optional = organAccountRepository.getByOgnIdAndReferMobile(ognId, mobile);
         if (!optional.isPresent()) {
             log.error("[随堂评论]新增随堂评论失败，机构用户不存在。");
             return ResultDo.build(MessageCode.ERROR_ORGAN_ACCOUNT_NOT_EXISTS);
         }
         Long ognAccountId = optional.get().getId();
-        EpMemberChildCommentPo po = new EpMemberChildCommentPo();
-        po.setChildId(childId);
-        po.setOgnId(ognId);
-        po.setCourseId(courseId);
-        po.setClassId(classId);
-        po.setClassCatalogId(catalogId);
-        po.setType(EpMemberChildCommentType.launch);
-        po.setContent(content);
-        po.setOgnAccountId(ognAccountId);
-        log.error("[随堂评论]新增随堂评论开始，po={}。", po);
-        memberChildCommentRepository.insert(po);
-        log.info("[随堂评论]新增随堂评论成功，id={}。", po.getId());
+        EpMemberChildCommentPo epMemberChildCommentPo = new EpMemberChildCommentPo();
+        epMemberChildCommentPo.setChildId(childId);
+        epMemberChildCommentPo.setOgnId(ognId);
+        epMemberChildCommentPo.setCourseId(courseId);
+        epMemberChildCommentPo.setClassId(classId);
+        epMemberChildCommentPo.setClassCatalogId(catalogId);
+        epMemberChildCommentPo.setType(EpMemberChildCommentType.launch);
+        epMemberChildCommentPo.setContent(content);
+        epMemberChildCommentPo.setOgnAccountId(ognAccountId);
+        log.error("[随堂评论]新增随堂评论开始，epMemberChildCommentPo={},tagIds={}。", epMemberChildCommentPo, tagIds);
+        memberChildCommentRepository.insert(epMemberChildCommentPo);
+        List<EpMemberChildTagPo> insertPos = Lists.newArrayList();
+        if (CollectionsTools.isNotEmpty(tagIds)) {
+            tagIds.forEach(p -> {
+                EpMemberChildTagPo po = new EpMemberChildTagPo();
+                po.setChildId(childId);
+                po.setOgnId(ognId);
+                po.setCourseId(courseId);
+                po.setClassId(classId);
+                po.setClassCatalogId(catalogId);
+                po.setTagId(p);
+                insertPos.add(po);
+            });
+        }
+        memberChildTagRepository.insert(insertPos);
+        log.info("[评论]修改评论成功，评论id={}。");
+        log.info("[随堂评论]新增随堂评论成功，id={}。", epMemberChildCommentPo.getId());
+        return ResultDo.build();
+    }
+
+    /**
+     * 由商户账号后台创建随堂评论
+     *
+     * @param childId
+     * @param ognId
+     * @param courseId
+     * @param classId
+     * @param catalogId
+     * @param content
+     * @param userId
+     * @return
+     */
+    @Transactional(rollbackFor = Exception.class)
+    public ResultDo createCommentLaunchByOgn(Long childId, Long ognId, Long courseId, Long classId, Long catalogId, String content, Long userId, List<Long> tagIds) {
+
+        EpMemberChildCommentPo epMemberChildCommentPo = new EpMemberChildCommentPo();
+        epMemberChildCommentPo.setChildId(childId);
+        epMemberChildCommentPo.setOgnId(ognId);
+        epMemberChildCommentPo.setCourseId(courseId);
+        epMemberChildCommentPo.setClassId(classId);
+        epMemberChildCommentPo.setClassCatalogId(catalogId);
+        epMemberChildCommentPo.setType(EpMemberChildCommentType.launch);
+        epMemberChildCommentPo.setContent(content);
+        epMemberChildCommentPo.setOgnAccountId(userId);
+        log.error("[随堂评论]新增随堂评论开始，epMemberChildCommentPo={},tagIds={}。", epMemberChildCommentPo, tagIds);
+        memberChildCommentRepository.insert(epMemberChildCommentPo);
+        List<EpMemberChildTagPo> insertPos = Lists.newArrayList();
+        if (CollectionsTools.isNotEmpty(tagIds)) {
+            tagIds.forEach(p -> {
+                EpMemberChildTagPo po = new EpMemberChildTagPo();
+                po.setChildId(childId);
+                po.setOgnId(ognId);
+                po.setCourseId(courseId);
+                po.setClassId(classId);
+                po.setClassCatalogId(catalogId);
+                po.setTagId(p);
+                insertPos.add(po);
+            });
+            memberChildTagRepository.insert(insertPos);
+        }
+
+        log.info("[随堂评论]新增随堂评论成功，id={}。", epMemberChildCommentPo.getId());
         return ResultDo.build();
     }
 }
